+++ src/emu/arithmetic/adi.rs
@@ -8,13 +8,14 @@ use super::get_flags;
 pub fn execute_adi(state: &mut State, adi: ADI) {
     match adi {
         ADI::AddImmediate { data } => {
-            let lhs = state.get_register(&Register::A) as i8;
-            let rhs = data as i8;
+            // Only subtraction uses two's complement, so these are unsigned.
+            let lhs = state.get_register(&Register::A);
+            let rhs = data;
 
             let (result, carried) = lhs.overflowing_add(rhs);
             let flags = get_flags(lhs, rhs, result, carried);
 
-            state.set_register(&Register::A, result as u8);
+            state.set_register(&Register::A, result);
             state.alu.flags = flags;
         }
     }
