+++ src/emu/arithmetic/inr.rs
@@ -8,13 +8,14 @@ use super::get_flags;
 pub fn execute_inr(state: &mut State, inr: INR) {
     match inr {
         INR::IncrementRegister { r } => {
-            let lhs = state.get_register(&r) as i8;
-            let rhs = 1_i8;
+            // Only subtraction uses two's complement, so these are unsigned.
+            let lhs = state.get_register(&r);
+            let rhs = 1_u8;
 
             let (result, carried) = lhs.overflowing_add(rhs);
             let flags = get_flags(lhs, rhs, result, carried);
 
-            state.set_register(&r, result as u8);
+            state.set_register(&r, result);
             state.alu.flags = Flags {
                 zero: flags.zero,
                 carry: state.alu.flags.carry, // Carry flag isn't affected.
@@ -25,13 +26,14 @@ pub fn execute_inr(state: &mut State, inr: INR) {
         }
         INR::IncrementMemory => {
             let address = state.get_register_pair(&RegisterPair::HL);
-            let lhs = state.get_memory(address) as i8;
-            let rhs = 1_i8;
+            // Only subtraction uses two's complement, so these are unsigned.
+            let lhs = state.get_memory(address);
+            let rhs = 1_u8;
 
             let (result, carried) = lhs.overflowing_add(rhs);
             let flags = get_flags(lhs, rhs, result, carried);
 
-            state.set_memory(address, result as u8);
+            state.set_memory(address, result);
             state.alu.flags = Flags {
                 zero: flags.zero,
                 carry: state.alu.flags.carry, // Carry flag isn't affected.
