+++ src/emu/logical/cmp.rs
@@ -9,8 +9,8 @@ use crate::{
 pub fn execute_cmp(state: &mut State, cmp: CMP) {
     match cmp {
         CMP::CompareRegister { r } => {
-            let lhs = state.get_register(&Register::A) as i8;
-            let rhs = state.get_register(&r) as i8;
+            let lhs = state.get_register(&Register::A);
+            let rhs = state.get_register(&r);
 
             let (result, carried) = lhs.overflowing_sub(rhs);
             let flags = arithmetic::get_flags(lhs, rhs, result, carried);
@@ -20,8 +20,8 @@ pub fn execute_cmp(state: &mut State, cmp: CMP) {
         }
         CMP::CompareMemory => {
             let address = state.get_register_pair(&RegisterPair::HL);
-            let lhs = state.get_register(&Register::A) as i8;
-            let rhs = state.get_memory(address) as i8;
+            let lhs = state.get_register(&Register::A);
+            let rhs = state.get_memory(address);
 
             let (result, carried) = lhs.overflowing_sub(rhs);
             let flags = arithmetic::get_flags(lhs, rhs, result, carried);
