+++ src/emu/logical/cpi.rs
@@ -6,8 +6,8 @@ use crate::{
 pub fn execute_cpi(state: &mut State, cpi: CPI) {
     match cpi {
         CPI::CompareImmediate { data } => {
-            let lhs = state.get_register(&Register::A) as i8;
-            let rhs = data as i8;
+            let lhs = state.get_register(&Register::A);
+            let rhs = data;
 
             let (result, carried) = lhs.overflowing_sub(rhs);
             let flags = arithmetic::get_flags(lhs, rhs, result, carried);
